requirements.txt
streamlit
torch>=2.0.0
transformers>=4.30.0
opencv-python
Pillow
numpy
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git
git+https://github.com/openai/CLIP.git